update docs and test
